Target lines and click detection (#10)
Added target lines that are situated between the turret and the target and that forme an interrupted line between them. Also fixed click detection so that firing a shell happens strictly when you click LMB, not when you press LMB and move the mouse around, like before. Also changed the turret image so that now the turret rotattes ot around the center of the image but around around the point where the gun starts, like in a real tank.

* Add files via upload

* Delete logfile.txt

* Delete Bullet_1.png

* Delete Sand_1.png

* Delete Tank_1.png

* Delete GunTurret_1.png

* Delete tank_target.png

* Delete project.greenfoot

* Delete tank_moving.wav

* Delete tank_moving_1.wav

* Delete README.TXT

* Delete Shell.class

* Delete Shell.ctxt

* Delete Shell.java

* Delete Tank.class

* Delete Tank.ctxt

* Delete Tank.java

* Delete TankWorld.class

* Delete TankWorld.ctxt

* Delete TankWorld.java

* Delete Target.class

* Delete Target.ctxt

* Delete Target.java

* Delete Turret.class

* Delete Turret.ctxt

* Delete Turret.java

* Add files via upload