First level (#19)
Completed the first level. Now, a mesage appears telling the player the level and the number of enemy tanks, there UI elements on the screen during the game telling you about the number of enemy tanks and player lives left. If you win the mission, a message will be displayed and the next level will be loaded. If you die, the current level will be reloaded and you will lose one life. If you have no more lives left, the game stops.

* Update BrownTank.java

* Add files via upload

* Update PlayerTank.java

* Update PlayerTurret.java

* Update PlayerTurret.java

* Update PlayerTurret.java

* Delete PlayerTurret.java

* Add files via upload

* Update Shell.java

* Update Tank.java

* Update TankWorld.java

* Update Turret.java

* Delete project.greenfoot

* Add files via upload

* Add files via upload