Graph source point changes (#44)
Changed the way the source point in the shortest path algorithm is not a newly created one, but one already in the graph whose distance is set to 0. This source point will not be put into the path that is returned to make the tank movement more natural. Also changed a variable in the TurquoiseTank class to static as it should have been.

* Update Graph.java

* Update Graph.java

* Update GraphPoint.java

* Update TurquoiseTank.java