Turning near walls and bug fixing (#48)
Made mobile enemy tanks not move while turning if by doing so they would enter a wall. Also fixed a bug that would throw a NullPointerException very rarely when the enemy tank generated a new path to follow. Added images for the turquoise tank and turret that I forgot to add earlier.

* Update MobileEnemyTank.java

* Update Graph.java

* Update Tank.java

* Add files via upload