Destroyable walls and rocket shells (#51)
Made the old type of wall impervious to landmines and added a new type of wall with a different image that is destroyable. Made turquoise turrets fire a new type of shell, a rocket shell, which moves faster but does not bounce. Also fixed a bug that would make the turquoise tanks drive through walls sometimes after the level was reloaded because the player lost.

* Add files via upload

* Delete project.greenfoot

* Add files via upload

* Update EnemyTurret.java

* Update LandMine.java

* Update PlayerTurret.java

* Update Shell.java

* Update TankWorld.java

* Update TurquoiseTank.java

* Update TurquoiseTurret.java

* Update Turret.java

* Delete EnemyTurret.class

* Delete LandMine.class

* Delete PlayerTank.class

* Delete PlayerTurret.class

* Delete Shell.class

* Delete TankWorld.class

* Delete TurquoiseTank.class

* Delete TurquoiseTurret.class

* Delete Turret.class

* Add files via upload

* Add files via upload

* Add files via upload